rename base class
