+++ pages/order_page.py
@@ -2,7 +2,7 @@ import allure
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from base.base_class import BaseClass
+from base.base_page import BaseClass
 from src import data
