+++ pages/registration_page.py
@@ -1,7 +1,7 @@
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from base.base_class import BaseClass
+from base.base_page import BaseClass
 from src import urls
